#390: CategoryButtonCollectionViewCell > CategoryTagCollectionViewCell 네이밍 변경
